Move search/filter to components/Filter
